Rename project and taskodoro core module
