+++ .idea/.name
@@ -1,1 +1,1 @@
-Taskodoro
+Taskodoro_App
