+++ apps/android/app/build.gradle.kts
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2022 Felipe Joglar
+ *    Copyright 2023 Felipe Joglar
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -84,7 +84,7 @@ android {
 dependencies {
     coreLibraryDesugaring(libs.android.desugar)
 
-    implementation(projects.taskodoroCore)
+    implementation(projects.taskodoro)
 
     implementation(libs.bundles.androidx.compose)
     implementation(libs.androidx.activity.compose)
