--- taskodoro-core/src/nativeMain/kotlin/com/taskodoro/storage/db/DriverFactory.kt
@@ -1,27 +1,0 @@
-/*
- *    Copyright 2022 Felipe Joglar
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-package com.taskodoro.storage.db
-
-import com.squareup.sqldelight.db.SqlDriver
-import com.squareup.sqldelight.drivers.native.NativeSqliteDriver
-
-actual class DriverFactory {
-
-    actual fun createDriver(): SqlDriver {
-        return NativeSqliteDriver(TaskodoroDB.Schema, "TaskodoroDB")
-    }
-}
