+++ taskodoro/src/commonMain/kotlin/com/taskodoro/tasks/data/TaskRepository.kt
@@ -14,30 +14,14 @@
  *    limitations under the License.
  */
 
-pluginManagement {
-    repositories {
-        gradlePluginPortal()
-        google()
-        mavenCentral()
-    }
+package com.taskodoro.tasks.data
+
+import com.taskodoro.tasks.model.Task
+
+class TaskRepository(
+    private val localDataSource: TaskLocalDataSource,
+) {
+
+    fun getTasks(): List<Task> = localDataSource.getAllTasks()
 }
 
-rootProject.name = "Taskodoro_App"
-
-include(":apps:android:app")
-
-include(":taskodoro")
-
-enableFeaturePreview("VERSION_CATALOGS")
-enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
-
-dependencyResolutionManagement {
-    versionCatalogs {
-        create("config") {
-            from(files("gradle/catalogs/config.versions.toml"))
-        }
-        create("libs") {
-            from(files("gradle/catalogs/libs.versions.toml"))
-        }
-    }
-}
