+++ taskodoro/src/nativeTest/kotlin/com/taskodoro/storage/db/DriverFactory.kt
@@ -1,0 +1,38 @@
+/*
+ *    Copyright 2023 Felipe Joglar
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package com.taskodoro.storage.db
+
+import co.touchlab.sqliter.DatabaseConfiguration
+import com.squareup.sqldelight.db.SqlDriver
+import com.squareup.sqldelight.drivers.native.NativeSqliteDriver
+import com.squareup.sqldelight.drivers.native.wrapConnection
+
+actual class DriverFactory {
+    actual fun createDriver(): SqlDriver {
+        val schema = TaskodoroDB.Schema
+        return NativeSqliteDriver(
+            DatabaseConfiguration(
+                name = "TaskodoroDB",
+                version = schema.version,
+                create = { connection ->
+                    wrapConnection(connection) { schema.create(it) }
+                },
+                inMemory = true
+            )
+        )
+    }
+}
